Rename the utf8_array module to _utf8_array
Exporting blosc2.utf8_array() as the public Utf8Array constructor shadowed
the module of the same name.  `from blosc2.utf8_array import X` still
resolved (the import machinery finds the submodule), but attribute-path
lookups landed on the function, which broke two monkeypatch call sites in
the tests and would have confused anyone reading `blosc2.utf8_array` in a
traceback.

The module was always internal -- every reference to it is inside blosc2
or its tests -- so the leading underscore states that and frees the plain
name for the constructor.  The monkeypatch sites go back to the plain
attribute-path form.

No public API change: blosc2.utf8_array (function), blosc2.Utf8Array and
`from blosc2 import ...` are unaffected.
